server/compliance/pyproject.toml
[project]
name = "compliance-mcp"
version = "0.1.0"
description = "Cloud Security Compliance MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.1",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.8.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "google-cloud-cloudsecuritycompliance>=0.0.0",
    "uvicorn[standard]",
]

[project.scripts]
compliance_mcp = "compliance_mcp:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 